_tipsの総因数分解のバグ修正

+++ AtCoder/_tips.cpp
@@ -61,7 +61,7 @@ void tip_facotization()
 {
   ll n;
   map<ll, ll> primes;
-  int factor = 2;
+  ll factor = 2;
   while(factor * factor <= n)
   {
     if (n % factor == 0)
